dao layer

+++ backend/payment/src/main/java/com/dbs/payment/dao/CustomerRepository.java
@@ -1,9 +1,11 @@
 package com.dbs.payment.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dbs.payment.model.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer>{
+public interface CustomerRepository extends JpaRepository<Customer, BigInteger>{
 
 }
